sql: move expression transformations to dedicated module
This makes it a bit easier to see an individual transform, since it
isn't all mixed up with other random expr-related code.
